backend/pyproject.toml
[project]
name = "intel-hub-backend"
version = "0.1.0"
description = "Personal Intelligence Hub backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.31.0",
  "sqlalchemy>=2.0.36",
  "alembic>=1.13.3",
  "psycopg[binary]>=3.2.3",
  "redis>=5.2.0",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.6.1",
  "httpx>=0.27.2",
  "feedparser>=6.0.11",
  "tenacity>=9.0.0",
  "apscheduler>=3.10.4",
  "python-telegram-bot>=21.7",
  "python-socketio>=5.11.4",
  "python-dateutil>=2.9.0.post0",
  "prometheus-client>=0.21.0",
  "structlog>=24.4.0",
]

[dependency-groups]
dev = [
  "pytest>=8.3.3",
  "pytest-cov>=5.0.0",
  "pytest-mock>=3.14.0",
  "ruff>=0.7.4",
]

[tool.uv]
default-groups = ["dev"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = ["E501", "B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
